+++ instagram-app/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "bootstrap": "^4.3.1",
+    "fuzzy": "^0.1.3",
     "jquery": "^3.4.1",
     "node-sass": "^4.12.0",
     "popper": "^1.0.1",
