+++ instagram-app/src/components/Login/Login.scss
@@ -1,0 +1,14 @@
+.loginBtn{
+    border:1px black solid !important;
+    border-radius: 10px !important;
+    margin:1rem !important;
+}
+fieldset{
+    border: lightgray 3px groove !important;
+    padding:.5rem !important;
+}
+legend{
+    width:auto !important;
+    margin-left:1rem;
+    color:red !important;
+}
